Moved plotting functions into a separate file, along with needed imports.
